// src/app/module/parts/model/model.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Fabricator } from './fabricator';


export interface Model {
    fabricator?: Fabricator;
    id?: number;
    name?: string;
}